refactor: Keep monitoring logic in the same file
